video: tegra: nvmap: Fix memory leak
When a handle is duplicated as RO, then even after calling
NvRmMemHandleFree API, the duplicated handle is not getting freed due to
incorrect refcount on the dma_buf. Fix it by decrementing an extra
dma_buf refcount.
The failure flow:
 - Create RW handle
 - Duplicate handle with RO
 - Free RO handle
 - Free RW handle

__nvmap_make_dmabuf creates a dmabuf and it would already have the
counter value set. In case of RO duplication, nvmap_duplicate_handle
would increment it, which is not necessary. But in case of creating
handle from RO FD, nvmap_duplicate_handle would be called, and if handle
reference is not present then reference is created and in that case it
is necessary to increment the dma_buf counter, as the counter is
maintained in such a way that, for every new reference creation, the
dma_buf counter should be incremented, so that while destroying the
reference the dma_buf counter decrement by 1.

Bug 3479904

Change-Id: Ib4725dc5edd38018b56b93852a436693c3218a00
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2633981
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit