video: tegra: nvmap: Fix double free case
nvmap_altfree already free the pages pointer but does not set it to NULL
If the execution jumps to out label then it checks whether pages pointer
is NULL or not, if not then it tries it free it again. In order to
avoid double free call, set pages pointer to NULL after call to
nvmap_altfree.

CID 10127996

Bug 3461002

Change-Id: Iced862d547f0fa288303f4c9bda254728e7afc09
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2658060
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit