soc: tegra: pmc: fix dual edge triggered wakes
When a wake event is defined to be triggered on both positive
and negative edge of the wake signal, it is crucial to know
the current state of the signal when going into suspend.
The intended way to obtain the current state of the wake
signals is to read the WAKE_AOWAKE_SW_STATUS register,
which should contains the raw state of the wake signals.

However, this register is edge triggered, an edge will not
be generated for signals that are already asserted prior to
the assertion of WAKE_LATCH_SW.

To workaround this, change the polarity of the wake level
from '0' to '1' while latching the signals,
as this will generate an edge for signals that are set to '1'.

Bug 1737058
Bug 3474189
Bug 200779609

Change-Id: Ie20fd0eb148ff4e87c935b4574a25b785fcbdaf0
Signed-off-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-on: http://git-master/r/1182638
(cherry picked from commit a4649120a9d83381504f9264d6f13e8f564724ba)
Signed-off-by: Petlozu Pravareshwar <petlozup@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2659940
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Kartik . <kkartik@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit