video: tegra: nvmap: Fix stack frame size warning
When building NVMAP for Linux kernels where the CONFIG_FRAME_WARN is
set to 1024, the following build error is observed ...

 drivers/video/tegra/nvmap/nvmap_ioctl.c: In function ‘find_free_cma_mem’:
 drivers/video/tegra/nvmap/nvmap_ioctl.c:1268:1: error: the frame size of
	1552 bytes is larger than 1024 bytes [-Werror=frame-larger-than=]

The problem is that find_free_cma_mem() function defines an array of
size 1536 which is greater than 1024. Rather than parsing the file
/proc/meminfo to determine the amount of free CMA space, call the
function global_zone_page_state() to get the number of free CMA pages.
Please note that global_zone_page_state() is the function that is called
by when reading /proc/meminfo.

Change-Id: Ib2ca071e6e7227f3eda34c332e3123f0e3e49ccc
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2663538
Reviewed-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Martin Radev <mradev@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit