video: tegra: nvmap: Fix trace use-after-free
Enabling the 'nvmap_destroy_handle' exposes a use-after-free issue in
the function _nvmap_handle_free(). The problem is that after freeing
the carveout, the carveout pointer is still valid and so when the trace
function is called, we attempt to deference the handle->carveout
pointer. Fix this by clearing the carveout pointer after it has been
freed.

Bug 3552534

Change-Id: I13de0515cae188d88b187957dba15ad7d35f24c1
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2675331
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>