soc: t234: declare sysram as device memory
The IPC memory carveout used by CPU to communicate with BPMP-FW is 8kB
in size. The carveout is followed by other secure carveouts that CPU
is not permitted to access. When 64kB page size is used, the CPU is
allowed to make speculative reads that can land outside of the 8kB
carveout and generate a SError.

Mark sysram as device memory with 'no-memory-wc' configuration
property to prevent speculative reads. This change has minor
performance implications to IPC communication (measured as 10%
increase in round trip delay) and requires CPU to not perform
unaligned accesses to the memory area.

Bug 3553190
Bug 3473388
Bug 3469696

Change-Id: If7a5b4b0affd1fcc026acf4d08952ca04cf85a16
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t23x/+/2676819
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>