video: tegra: nvmap: Don't treat ivm as reserved mem carveouts
Making NvMap as LKM for 5.10 is leading to following unresolved symbol:
- of_get_flat_dt_prop
This function is used to get ivm property from the reserved memory fdt
node for ivm. This function works on flattend DT blob and it is a non
exported function. We can not simply replace this with the function like
of_find_compatible_node which works on the unflattened device tree.
on TOT, The ivm carveout is managed as reserved memory carveout but
actually the reserved memory framwork is not allocating the memory for
it i.e. the tegra-hv driver allocate the memory for the ivm carvout and
assigned with the help of tegra_hv_mempool_reserve function.

As the memory region for the carveout is not getting reserved through
the reserved mem, don't treat ivm as reserved mem carveouts. Move the
ivm carveout node out of reserved-memory node in DT. In NvMap driver,
during nvmap_init, parse the DT for the ivm carveout nodes, instead of
using the reserved mem APIs. With this, we can get rid of use of the
of_get_flat_dt_prop function.

Bug 200722684

Change-Id: I69432dfc5423bfa20688018f711282e5b31bb0d3
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2680105
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>