pinctrl: tegra: Add support for setting the GPIO direction
The Tegra pinctrl driver does not have support for setting the pin
direction when requesting a GPIO. This means that if the default pin
configuration does not have the input-enable and tristate (for
disabling the output) bits set appropriately, then GPIOs may not
function as expected. Add a helper function that configures the
input-enable and tristate bit as necessary for a GPIO.

Also, remove num_ranges functionality from gpio driver.
so the pinctrl driver will use the gpio-ranges property from DT.

Bug 3499398

Change-Id: I393cfb2036f339ac886c83898d2c5cd552542840
Signed-off-by: Prathamesh Shete <pshete@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2725833
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>