video: tegra: nvmap: Fix NVSCIIPC support
When building NVMAP as an external module, although NVMAP_CONFIG_SCIIPC
is enabled to build SCIIPC, because the compilation flag CONFIG_NVSCIIPC
is never enabled, the NVMAP driver is not compiled so that it actually
calls into the NVSCIIPC driver.

The Kconfig symbol CONFIG_NVSCIIPC is used to indicate if support for
the SCIIPC driver is enabled. Previously, if CONFIG_NVSCIIPC was
enabled, this would then enable NVMAP_CONFIG_SCIIPC which is the correct
way to handle this because we only want to enable NVMAP SCIIPC support
if the SCIIPC driver is enabled. Now we always enable NVMAP SCIIPC
support regardless of if the SCIIPC driver is enabled. Fix this by
partially reverting some of the recent changes made.

Bug 200753948

Change-Id: I7d4f926d09571e4a57e043f1a2ac6f673c45e139
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2689285
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit