misc: nvsciipc: Add support for upstream Linux
The NVSCIIPC driver is a simple driver that is not associated with
any physical hardware device. Device-tree is used to describe physical
hardware devices and given that no such device exists, there is no need
for device-tree to represent this device. For upstream Linux where this
device-tree node does not exist, we can simply register the device
within the driver when the driver is registered.

Bug 200753948

Change-Id: Ide73836d712f76b1e944c33fb97ac6f5af21b8ce
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2689799
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit