soc: t23x: kernel-dts: Add vpr to memory-region node
In case of nvmap as LKM, the memory-region list in tegra-carveouts
device tree node is parsed for determining reserved memory carveouts
that need to be created. Add vpr node phandle to memory-region so that
vpr carveout will be created.

Remove unncessary properties from vpr node and keep the status as
disabled. UEFI updates the reg and status of VPR node, based upon the
registers for vpr.

Bug 200722684

Change-Id: Ic291901058538dd310a44ece344244433fcd4650
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t23x/+/2691832
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Ketan Patil <ketanp@nvidia.com>
GVS: Gerrit_Virtual_Submit