video: tegra: nvmap: Enable legacy init support
Legacy init support is needed in case of initializing vpr base, size in
case of backdoor boot, as UEFI is not present.
This legacy init support should be present even after making nvmap as a
loadable module. In case of kstable, we are not compiling tegra-vpr
driver hence skip this support only for kstable.

Bug 200722684

Change-Id: Id69ffd0cf62eece4d0098949c3c0dea74a746e2d
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2692618
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit