BACKPORT: cpufreq: tegra234: add support for T234
In upstream, support for T234 is added in the same file as T194.
In this patch, synched the downstream driver file for T194 to
the latest upstream file. Then backported changes that were required
for adding support for T234 in the same file as T194.

Upstream latest patch series:
  "https://lore.kernel.org/lkml/
  20220411030030.eqq4a2xgidexowzq@vireshk-i7/T/"

Bug 3625153

Change-Id: I0e9c805ec097f45b502f6803ed0a8215aa8dad90
Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2694544
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Sanjay Chandrashekara <sanjayc@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
GVS: Gerrit_Virtual_Submit