drivers: nvpps: support nvpps init in timer mode
support nvpps init in timer mode when 1PPS GPIO
pin is not provided/available in dt.
On DriveOrin, the PPS GPIO pin is not available
and currently the driver need PPS GPIO pin to
initialize in default GPIO mode.
This patch adds support to initialize the driver
in timer mode when GPIO pin is not found int dt

Bug 3595073
Jira ESDP-15159

Change-Id: I28a3acef39d809fb59d1bf9f89e332fae8d0d548
Signed-off-by: Sheetal Tigadoli <stigadoli@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2699790
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: David Tao <tehyut@nvidia.com>
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>
GVS: Gerrit_Virtual_Submit