VSE Driver: Generate random IV for encryption
This commit adds a command to VSE driver for generation of random IVs
for AES encryption. Not all AES encryptions require a random IV.
Driver does not know which ones do and therefore calls made to the
driver need to instruct it to generate random IVs when required.
Since the Crypto framework of Linux kernel does not provide any
convenient way to communicate this need to the driver, we have had to
make use of *iv* field in *struct skcipher_request* for this purpose.
When the first byte of the *iv* field in an encryption request is set to 1,
IV generation is required. If the first byte has any other value, IV
generation is not required. If the *iv* field is NULL, the encryption
request is considered invalid and an error is returned.

JIRA SSV-8759

Change-Id: I8aaab76a8a78f650fa94a9619bb1b76abaf7bd7e
Signed-off-by: Rounak Agarwal <rounaka@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2701477
Reviewed-by: Leo Chiu <lchiu@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
GVS: Gerrit_Virtual_Submit