drivers: pva: buffer support offset & size
userspace expects to pin buffers by offse and size.
buffers with different sizes or offsets relative to
backing dma_buf will be pinned once more and offset
and size information will be stored in the driver.

Future optimization is to track if dma_buf is pinned
and not pin it more than once.

Bug 3587671
Bug 3345117

Signed-off-by: Omar Nemri <onemri@nvidia.com>
Change-Id: I263393c3d6000e8458e36c40e931eae2f0c8ce80
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2702126
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
GVS: Gerrit_Virtual_Submit